Create intersect.py
This script takes two files that are on a local hard drive and compares them. One file has the list of companies that are in the Dow Jones 30 and one has a list of the companies that are in the Nasdaq 100.

A future version will read the information from the web.